✨ feat: float-only meter input; reset resampler state on meter reset
The meter now accepts only float32/float64 numpy arrays and works
internally in f32 like the reference P.56; integer WAV files are
normalized to floats at the reader boundary (soundfile divides by
127/32767/8388607/2147483647 for 8/16/24/32-bit files). bit_depth stays
as the histogram-span parameter only.

Also fixes reset(): it now clears the internal <16 kHz resampler state
so finish() -> reset() -> process_block() works again.

+++ tests/test_prefilter.py
@@ -33,11 +33,11 @@ def test_invalid_sample_rate() -> None:
 def test_process_returns_float32() -> None:
     """process() returns a new float32 array; input is not modified."""
     pf = PreFilter("NB", 16_000)
-    x = np.zeros(1024, dtype=np.int16)
+    x = np.zeros(1024, dtype=np.float32)
     out = pf.process(x)
     assert out.dtype == np.float32
     assert out.shape == (1024,)
-    np.testing.assert_array_equal(x, np.zeros(1024, dtype=np.int16))
+    np.testing.assert_array_equal(x, np.zeros(1024, dtype=np.float32))
 
 
 def test_streaming_matches_one_shot() -> None:
